fixed floating point conversion in relu tester

+++ samples/eltwise/eltwise_unary_relu.c
@@ -29,7 +29,7 @@ void relu_fwd_gold(const libxsmm_blasint M, const libxsmm_blasint N, const libxs
         } else if ( type == 1 ) {
           f_out[(j*ldo) + i] = ( f_in[(j*ldi) + i] < 0.0f ) ? alpha*f_in[(j*ldi) + i] : f_in[(j*ldi) + i];
         } else if ( type == 2 ) {
-          f_out[(j*ldo) + i] = ( f_in[(j*ldi) + i] < 0.0f ) ? alpha * ((float)expf(f_in[(j*ldi) + i])-1.0) : f_in[(j*ldi) + i];
+          f_out[(j*ldo) + i] = ( f_in[(j*ldi) + i] < 0.0f ) ? alpha * (expf(f_in[(j*ldi) + i])-1.0f) : f_in[(j*ldi) + i];
         }
         if ( type != 2) {
           out_mask[(j*ldo_mask) + i/8] |= (unsigned char)(( f_in[(j*ldi) + i] < 0.0f ) ? 0x0 : (1 << (i%8)) );
@@ -48,7 +48,7 @@ void relu_fwd_gold(const libxsmm_blasint M, const libxsmm_blasint N, const libxs
         } else if ( type == 1 ) {
           out_value = ( in_value < 0.0f ) ? alpha*in_value : in_value;
         } else if ( type == 2 ) {
-          out_value = ( in_value < 0.0f ) ? alpha*((float)expf(in_value)-1.0) : in_value;
+          out_value = ( in_value < 0.0f ) ? alpha*(expf(in_value)-1.0f) : in_value;
         }
         if ( type != 2) {
           out_mask[(j*ldo_mask) + i/8] |= (unsigned char)(( in_value < 0.0f ) ? 0x0 : (1 << (i%8)) );
@@ -67,7 +67,7 @@ void relu_fwd_gold(const libxsmm_blasint M, const libxsmm_blasint N, const libxs
         } else if ( type == 1 ) {
           out_value = ( f_in[(j*ldi) + i] < 0.0f ) ? alpha*f_in[(j*ldi) + i] : f_in[(j*ldi) + i];
         } else if ( type == 2 ) {
-          out_value = ( f_in[(j*ldi) + i] < 0.0f ) ? alpha * ((float)expf(f_in[(j*ldi) + i])-1.0) : f_in[(j*ldi) + i];
+          out_value = ( f_in[(j*ldi) + i] < 0.0f ) ? alpha * (expf(f_in[(j*ldi) + i])-1.0f) : f_in[(j*ldi) + i];
         }
         if ( type != 2) {
           out_mask[(j*ldo_mask) + i/8] |= (unsigned char)(( f_in[(j*ldi) + i] < 0.0f ) ? 0x0 : (1 << (i%8)) );
@@ -87,7 +87,7 @@ void relu_fwd_gold(const libxsmm_blasint M, const libxsmm_blasint N, const libxs
         } else if ( type == 1 ) {
           f_out[(j*ldo) + i] = ( in_value < 0.0f ) ? alpha*in_value : in_value;
         } else if ( type == 2 ) {
-          f_out[(j*ldo) + i] = ( in_value < 0.0f ) ? alpha*((float)expf(in_value)-1.0) : in_value;
+          f_out[(j*ldo) + i] = ( in_value < 0.0f ) ? alpha*(expf(in_value)-1.0f) : in_value;
         }
         if ( type != 2) {
           out_mask[(j*ldo_mask) + i/8] |= (unsigned char)(( in_value < 0.0f ) ? 0x0 : (1 << (i%8)) );
